[Chore] API DTO Nested Type으로 변경 및 공통된 DTO는 NameSpace를 통해 분리
